requirements.txt
torch
torchvision
torch-lightning
wandb
